Versión 2.11.11
- Se generaron los pedido a proveedores y se aprobaron y enviaron los mail manualmente.
- Se arreglaron la lista de asistencia y crearon el registro de asistencia.
- Se modificaron los atributos de la tabla asistencia el estado y creo un nuevo atributo llamado tipoAsistencia en el backend.
- Se envían los enlaces de registrar asistencia con éxito usando la herramienta ngrok.
- Con la herramienta ngrok se pudo registrar asistencia desde el movil.
- Se crearon las tabla de consumo, y esta habilitado para la exportación de CSV y la generación de pdf.
- Para la generación de pdf se usaron librería como jspdf, jspdf-autotable, pdfkit.

+++ Comedor/client/src/pages/cocinera/CocineraInventario.jsx
@@ -70,9 +70,9 @@ const CocineraInventario = () => {
     const alertas = [];
 
     inventariosData.forEach((inv) => {
-      const cantidad = parseFloat(inv.cantidadActual);
-      const nivelMinimo = parseFloat(inv.nivelMinimoAlerta);
-      const stockMaximo = parseFloat(inv.stockMaximo) || nivelMinimo; // Fallback si no hay stockMaximo
+      const cantidad = parseInt(inv.cantidadActual);
+      const nivelMinimo = parseInt(inv.nivelMinimoAlerta);
+      const stockMaximo = parseInt(inv.stockMaximo) || nivelMinimo; // Fallback si no hay stockMaximo
 
       // Alerta si: stock actual <= stock mínimo de alerta
       if (cantidad <= nivelMinimo) {
